chore: move comparadise app into app directory
