+++ action/run.test.ts
@@ -4,6 +4,7 @@ import { getInput, getMultilineInput, setFailed } from '@actions/core';
 import { octokit } from './octokit';
 import { sync } from 'glob';
 import { VISUAL_REGRESSION_CONTEXT } from '../constants';
+import { expect } from '@jest/globals';
 
 jest.mock('glob');
 jest.mock('@actions/core');
