+++ app/backend/allNonVisualChecksHavePassed.test.ts
@@ -1,6 +1,7 @@
 import { allNonVisualChecksHavePassed } from './allNonVisualChecksHavePassed';
 import { getOctokit } from './getOctokit';
-import { VISUAL_REGRESSION_CONTEXT } from '../constants';
+import { VISUAL_REGRESSION_CONTEXT } from '../../constants';
+import { expect } from '@jest/globals';
 
 jest.mock('./getOctokit');
