+++ app/backend/getBase64StringFromS3.ts
