+++ app/backend/getGroupedImages.ts
