+++ app/backend/updateBaseImagesInS3.test.ts
@@ -2,6 +2,7 @@ import { S3Client } from './s3Client';
 import { filterNewImages, replaceImagesInS3, getBaseImagePaths, updateBaseImagesInS3 } from './updateBaseImagesInS3';
 import { BASE_IMAGES_DIRECTORY } from './constants';
 import { allNonVisualChecksHavePassed } from './allNonVisualChecksHavePassed';
+import { expect } from '@jest/globals';
 
 jest.mock('./allNonVisualChecksHavePassed');
 jest.mock('./s3Client');
