+++ app/backend/updateBaseImagesInS3.ts
