+++ app/frontend/components/image-views.tsx
