+++ app/frontend/components/landing-page.tsx
