+++ app/frontend/components/main-page.tsx
