+++ app/frontend/components/update-images-button.tsx
