+++ app/frontend/components/view-toggle.tsx
