+++ app/frontend/index.html
@@ -2,13 +2,13 @@
 <html lang="en">
   <head>
     <meta charset="UTF-8" />
-    <link rel="icon" type="image/svg+xml" href="/frontend/resources/Island.svg" />
+    <link rel="icon" type="image/svg+xml" href="./resources/Island.svg" />
     <link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Roboto:300,400,500,700&display=swap" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Comparadise</title>
   </head>
   <body>
     <div id="root"></div>
-    <script type="module" src="/frontend/main.tsx"></script>
+    <script type="module" src="./main.tsx"></script>
   </body>
 </html>
