+++ app/frontend/providers/base-image-state-provider.tsx
