+++ tsconfig.json
@@ -17,7 +17,5 @@
     "strict": true,
     "noEmit": true
   },
-  "include": ["action", "backend", "constants", "frontend"],
-  "exclude": ["dist", "node_modules"],
   "references": [{ "path": "./tsconfig.node.json" }]
 }
