+++ vite.config.ts
@@ -3,6 +3,7 @@ import react from '@vitejs/plugin-react-swc';
 
 // https://vitejs.dev/config/
 export default defineConfig({
+  root: './app/frontend',
   plugins: [react()],
   server: {
     proxy: {
